+++ core/TreeNode.ts
@@ -2444,7 +2444,7 @@ class TreeNode extends AbstractNode {
     return str ? indent + str.replace(/\n/g, indent) : ""
   }
 
-  static getVersion = () => "41.0.0"
+  static getVersion = () => "41.1.0"
 
   static fromDisk(path: string): TreeNode {
     const format = this._getFileFormat(path)