+++ releaseNotes.md
@@ -1,7 +1,8 @@
-edge / 2019-09-18
+41.1.0 / 2019-09-18
 ===================
 - New: error handling in TCF commands
 - Fix: uncaught command error displaying in Designer app
+- Fix: generateSimulatedData regression fix and tests
 
 41.0.0 / 2019-09-17
 ===================